feat(benchmark): restructure code files
