+++ tooling/benchmark.py
@@ -1,7 +1,4 @@
 import argparse
-import asyncio
-import json
-import wave
 from jiwer.transforms import (
     Compose,
     RemoveMultipleSpaces,
@@ -13,15 +10,20 @@ from jiwer.transforms import (
     ReduceToSingleSentence,
 )
 import tqdm
-import requests
 import os
 import time
 import jiwer
-from pathlib import Path
 import prettytable
-import websockets
 
-DATA_BASE_PATH = os.path.join(os.getcwd(), "data")
+from helpers.file_helper import (
+    load_file_list,
+    DATA_BASE_PATH,
+    get_expected_transcription,
+)
+
+from helpers.rest_helper import transcribe_file_rest
+from helpers.websocket_helper import transcribe_file_websocket
+
 
 transform_default = Compose(
     [
@@ -34,102 +36,6 @@ transform_default = Compose(
         ReduceToSingleSentence(),
     ]
 )
-
-
-async def read_wav_file(file_path):
-    with wave.open(file_path, "rb") as wav_file:
-        return wav_file.readframes(wav_file.getnframes())
-
-
-def load_file_list(size: str):
-    dirpath = os.path.join(DATA_BASE_PATH, size)
-    return [
-        os.path.join(dirpath, f)
-        for f in os.listdir(dirpath)
-        if os.path.isfile(os.path.join(dirpath, f)) and f.endswith(".wav")
-    ]
-
-
-def get_expected_transcription(path: str):
-    content = Path(f"{path}.txt").read_text()
-    content = content.replace("\n", "")
-    return transform_default(str(content))
-
-
-def await_transcription_finish(id: str, api_key: str):
-    is_finished = False
-    # TODO: add time cap
-    while not is_finished:
-        # Sleep for 10
-        # This MAY be a bit long but keeps possible impact on performance as low as possible
-        time.sleep(10)
-        r = requests.get(
-            "http://localhost:8393/transcriptions",
-            headers={"Authorization": api_key},
-        )
-        data = r.json()
-        for status in data:
-            if status["transcription_id"] != id:
-                continue
-            if status["status"] not in ["in_progress", "in_query"]:
-                is_finished = True
-
-
-# Wrap asyncio execution
-def transcribe_file_websocket(filepath: str) -> str:
-    loop = asyncio.get_event_loop()
-    res = loop.run_until_complete(asyncio.gather(__transcribe_file_websocket(filepath)))
-    print(res)
-    return res[0]
-
-
-# websockets is heavily integrated with asyncio...so we have to do this in an asyncio way
-async def __transcribe_file_websocket(filepath: str) -> str:
-    audio_data = await read_wav_file(filepath)
-
-    messages = []
-    try:
-        async with websockets.connect("ws://localhost:8394") as websocket_connection:
-            await websocket_connection.send(audio_data)
-            # Wait...the server needs a few seconds anyway
-            await asyncio.sleep(3)
-
-            while True:
-                try:
-                    message = await asyncio.wait_for(
-                        websocket_connection.recv(), timeout=15.0
-                    )
-                    messages.append(message)
-                except asyncio.TimeoutError:
-                    print("No message received for 5 seconds, sending 'eof'")
-                    await websocket_connection.close()
-                    break
-            assert len(messages) > 0
-    except websockets.exceptions.ConnectionClosedOK:
-        # This is the expected behaviour
-        pass
-    return json.loads(" ".join(messages))["text"]
-
-
-def transcribe_file_rest(filepath: str, api_key: str) -> str:
-    r = None
-    with open(filepath, "rb") as f:
-        r = requests.post(
-            "http://localhost:8393/transcriptions",
-            files={"file": f},
-            headers={"Authorization": api_key},
-        )
-    if r.status_code != 200:
-        print(r.status_code)
-        print("Something went wrong")
-        return ""
-    id = r.json()["transcription_id"]
-    await_transcription_finish(id, api_key)
-    transcription_result = requests.get(
-        f"http://localhost:8393/transcriptions/{id}",
-        headers={"Authorization": api_key},
-    )
-    return transcription_result.json()["transcript"]["text"]
 
 
 def benchmark(settings):
@@ -153,7 +59,7 @@ def benchmark(settings):
         elif settings.target == "websocket":
             transcription = transcribe_file_websocket(filepath)
         diff = time.time() - start_time
-        expected = get_expected_transcription(filepath)
+        expected = transform_default(get_expected_transcription(filepath))
         transcription = transform_default(transcription)
         if settings.debug:
             print(f"Expected: {expected}")
